requirement.txt

snowflake-connector-python==3.5.0
charset-normalizer==3.3.2
